verify if db updates already applied

+++ src/protected/db-updates.php
@@ -218,14 +218,20 @@ return [
     },
 
     'create saas tables' => function () use($conn) {
-      $conn->executeQuery("CREATE TABLE saas (id INT NOT NULL, name VARCHAR(255) NOT NULL, create_timestamp TIMESTAMP(0) WITHOUT TIME ZONE NOT NULL, status SMALLINT NOT NULL, agent_id INTEGER NOT NULL, PRIMARY KEY(id), url VARCHAR(255) NOT NULL, url_parent VARCHAR(255), slug VARCHAR(50) NOT NULL, namespace VARCHAR(50) NOT NULL);");
-      $conn->executeQuery("CREATE SEQUENCE saas_id_seq INCREMENT BY 1 MINVALUE 1 START 1;");
-      $conn->executeQuery("CREATE TABLE saas_meta ( object_id integer NOT NULL, key character varying(128) NOT NULL, value text, id integer NOT NULL);");
-      $conn->executeQuery("CREATE SEQUENCE saas_meta_id_seq INCREMENT BY 1 MINVALUE 1 START 1;");
-      $conn->executeQuery("ALTER TABLE ONLY saas_meta ADD CONSTRAINT saas_saas_meta_fk FOREIGN KEY (object_id) REFERENCES saas(id);");
+        if(__table_exists('saas')){
+            return true;
+        }
+        $conn->executeQuery("CREATE TABLE saas (id INT NOT NULL, name VARCHAR(255) NOT NULL, create_timestamp TIMESTAMP(0) WITHOUT TIME ZONE NOT NULL, status SMALLINT NOT NULL, agent_id INTEGER NOT NULL, PRIMARY KEY(id), url VARCHAR(255) NOT NULL, url_parent VARCHAR(255), slug VARCHAR(50) NOT NULL, namespace VARCHAR(50) NOT NULL);");
+        $conn->executeQuery("CREATE SEQUENCE saas_id_seq INCREMENT BY 1 MINVALUE 1 START 1;");
+        $conn->executeQuery("CREATE TABLE saas_meta ( object_id integer NOT NULL, key character varying(128) NOT NULL, value text, id integer NOT NULL);");
+        $conn->executeQuery("CREATE SEQUENCE saas_meta_id_seq INCREMENT BY 1 MINVALUE 1 START 1;");
+        $conn->executeQuery("ALTER TABLE ONLY saas_meta ADD CONSTRAINT saas_saas_meta_fk FOREIGN KEY (object_id) REFERENCES saas(id);");
     },
 
     'rename saas tables to subsite' => function () use($conn) {
+        if(__table_exists('subsite')){
+            return true;
+        }
         $conn->executeQuery("ALTER TABLE saas RENAME TO subsite");
         $conn->executeQuery("ALTER TABLE saas_meta RENAME TO subsite_meta");
         $conn->executeQuery("ALTER SEQUENCE saas_id_seq RENAME TO subsite_id_seq");
@@ -233,6 +239,9 @@ return [
     },
 
     'remove parent_url and add alias_url' => function () use($conn) {
+        if(__column_exists('subsite', 'alias_url')){
+            return true;
+        }
         $conn->executeQuery("ALTER TABLE subsite DROP COLUMN url_parent");
         $conn->executeQuery("ALTER TABLE subsite ADD COLUMN alias_url VARCHAR(255) DEFAULT NULL;");
 
@@ -243,6 +252,9 @@ return [
 
 
     'verified seal migration' => function () use($conn){
+        if($id = $conn->fetchColumn("SELECT id FROM seal WHERE id = 1")){
+            return true;
+        }
         $agent_id = $conn->fetchColumn("select profile_id
                     from usr
                     where id = (
@@ -266,14 +278,18 @@ return [
     },
 
     'create update timestamp entities' => function () use($conn) {
-    	$conn->executeQuery("ALTER TABLE agent ADD COLUMN update_timestamp TIMESTAMP(0) WITHOUT TIME ZONE;");
-    	$conn->executeQuery("ALTER TABLE space ADD COLUMN update_timestamp TIMESTAMP(0) WITHOUT TIME ZONE;");
-    	$conn->executeQuery("ALTER TABLE project ADD COLUMN update_timestamp TIMESTAMP(0) WITHOUT TIME ZONE;");
-    	$conn->executeQuery("ALTER TABLE event ADD COLUMN update_timestamp TIMESTAMP(0) WITHOUT TIME ZONE;");
-    	$conn->executeQuery("ALTER TABLE seal ADD COLUMN update_timestamp TIMESTAMP(0) WITHOUT TIME ZONE;");
+        if(!__column_exists('agent', 'update_timestamp')) $conn->executeQuery("ALTER TABLE agent ADD COLUMN update_timestamp TIMESTAMP(0) WITHOUT TIME ZONE;");
+    	if(!__column_exists('space', 'update_timestamp')) $conn->executeQuery("ALTER TABLE space ADD COLUMN update_timestamp TIMESTAMP(0) WITHOUT TIME ZONE;");
+    	if(!__column_exists('project', 'update_timestamp')) $conn->executeQuery("ALTER TABLE project ADD COLUMN update_timestamp TIMESTAMP(0) WITHOUT TIME ZONE;");
+    	if(!__column_exists('event', 'update_timestamp')) $conn->executeQuery("ALTER TABLE event ADD COLUMN update_timestamp TIMESTAMP(0) WITHOUT TIME ZONE;");
+    	if(!__column_exists('seal', 'update_timestamp')) $conn->executeQuery("ALTER TABLE seal ADD COLUMN update_timestamp TIMESTAMP(0) WITHOUT TIME ZONE;");
     },
 
     'alter table role add column subsite_id' => function () use($conn) {
+        if(__column_exists('role', 'subsite_id')){
+            return true;
+        }
+        
     	$conn->executeQuery("ALTER TABLE role DROP CONSTRAINT IF EXISTS role_user_fk;");
     	$conn->executeQuery("ALTER TABLE role DROP CONSTRAINT IF EXISTS role_unique;");
         $conn->executeQuery("ALTER TABLE role ADD subsite_id INT DEFAULT NULL;");
@@ -289,40 +305,61 @@ return [
     },
 
     'ADD columns subsite_id' => function () use($conn) {
-        $conn->executeQuery("ALTER TABLE space ADD subsite_id INT DEFAULT NULL;");
-        $conn->executeQuery("ALTER TABLE space ADD CONSTRAINT FK_2972C13AC79C849A FOREIGN KEY (subsite_id) REFERENCES subsite (id) NOT DEFERRABLE INITIALLY IMMEDIATE;");
-        $conn->executeQuery("CREATE INDEX IDX_2972C13AC79C849A ON space (subsite_id);");
+        if(!__column_exists('space', 'subsite_id')){
+            $conn->executeQuery("ALTER TABLE space ADD subsite_id INT DEFAULT NULL;");
+            $conn->executeQuery("ALTER TABLE space ADD CONSTRAINT FK_2972C13AC79C849A FOREIGN KEY (subsite_id) REFERENCES subsite (id) NOT DEFERRABLE INITIALLY IMMEDIATE;");
+            $conn->executeQuery("CREATE INDEX IDX_2972C13AC79C849A ON space (subsite_id);");
+        }
 
-        $conn->executeQuery("ALTER TABLE agent ADD subsite_id INT DEFAULT NULL;");
-        $conn->executeQuery("ALTER TABLE agent ADD CONSTRAINT FK_268B9C9DC79C849A FOREIGN KEY (subsite_id) REFERENCES subsite (id) NOT DEFERRABLE INITIALLY IMMEDIATE;");
-        $conn->executeQuery("CREATE INDEX IDX_268B9C9DC79C849A ON agent (subsite_id);");
+        if(!__column_exists('agent', 'subsite_id')){
+            $conn->executeQuery("ALTER TABLE agent ADD subsite_id INT DEFAULT NULL;");
+            $conn->executeQuery("ALTER TABLE agent ADD CONSTRAINT FK_268B9C9DC79C849A FOREIGN KEY (subsite_id) REFERENCES subsite (id) NOT DEFERRABLE INITIALLY IMMEDIATE;");
+            $conn->executeQuery("CREATE INDEX IDX_268B9C9DC79C849A ON agent (subsite_id);");
+        }
 
-        $conn->executeQuery("ALTER TABLE event ADD subsite_id INT DEFAULT NULL;");
-        $conn->executeQuery("ALTER TABLE event ADD CONSTRAINT FK_3BAE0AA7C79C849A FOREIGN KEY (subsite_id) REFERENCES subsite (id) NOT DEFERRABLE INITIALLY IMMEDIATE;");
-        $conn->executeQuery("CREATE INDEX IDX_3BAE0AA7C79C849A ON event (subsite_id);");
+        if(!__column_exists('event', 'subsite_id')){
+            $conn->executeQuery("ALTER TABLE event ADD subsite_id INT DEFAULT NULL;");
+            $conn->executeQuery("ALTER TABLE event ADD CONSTRAINT FK_3BAE0AA7C79C849A FOREIGN KEY (subsite_id) REFERENCES subsite (id) NOT DEFERRABLE INITIALLY IMMEDIATE;");
+            $conn->executeQuery("CREATE INDEX IDX_3BAE0AA7C79C849A ON event (subsite_id);");
+        }
 
-        $conn->executeQuery("ALTER TABLE project ADD subsite_id INT DEFAULT NULL;");
-        $conn->executeQuery("ALTER TABLE project ADD CONSTRAINT FK_2FB3D0EEC79C849A FOREIGN KEY (subsite_id) REFERENCES subsite (id) NOT DEFERRABLE INITIALLY IMMEDIATE;");
-        $conn->executeQuery("CREATE INDEX IDX_2FB3D0EEC79C849A ON project (subsite_id);");
+        if(!__column_exists('project', 'subsite_id')){
+            $conn->executeQuery("ALTER TABLE project ADD subsite_id INT DEFAULT NULL;");
+            $conn->executeQuery("ALTER TABLE project ADD CONSTRAINT FK_2FB3D0EEC79C849A FOREIGN KEY (subsite_id) REFERENCES subsite (id) NOT DEFERRABLE INITIALLY IMMEDIATE;");
+            $conn->executeQuery("CREATE INDEX IDX_2FB3D0EEC79C849A ON project (subsite_id);");
+        }
 
-        $conn->executeQuery("ALTER TABLE seal ADD subsite_id INT DEFAULT NULL;");
-        $conn->executeQuery("ALTER TABLE seal ADD CONSTRAINT FK_2E30AE30C79C849A FOREIGN KEY (subsite_id) REFERENCES subsite (id) NOT DEFERRABLE INITIALLY IMMEDIATE;");
-        $conn->executeQuery("CREATE INDEX IDX_2E30AE30C79C849A ON seal (subsite_id);");
+        if(!__column_exists('seal', 'subsite_id')){
+            $conn->executeQuery("ALTER TABLE seal ADD subsite_id INT DEFAULT NULL;");
+            $conn->executeQuery("ALTER TABLE seal ADD CONSTRAINT FK_2E30AE30C79C849A FOREIGN KEY (subsite_id) REFERENCES subsite (id) NOT DEFERRABLE INITIALLY IMMEDIATE;");
+            $conn->executeQuery("CREATE INDEX IDX_2E30AE30C79C849A ON seal (subsite_id);");
+        }
 
-        $conn->executeQuery("ALTER TABLE registration ADD subsite_id INT DEFAULT NULL;");
-        $conn->executeQuery("ALTER TABLE registration ADD CONSTRAINT FK_62A8A7A7C79C849A FOREIGN KEY (subsite_id) REFERENCES subsite (id) NOT DEFERRABLE INITIALLY IMMEDIATE;");
-        $conn->executeQuery("CREATE INDEX IDX_62A8A7A7C79C849A ON registration (subsite_id);");
+        if(!__column_exists('registration', 'subsite_id')){
+            $conn->executeQuery("ALTER TABLE registration ADD subsite_id INT DEFAULT NULL;");
+            $conn->executeQuery("ALTER TABLE registration ADD CONSTRAINT FK_62A8A7A7C79C849A FOREIGN KEY (subsite_id) REFERENCES subsite (id) NOT DEFERRABLE INITIALLY IMMEDIATE;");
+            $conn->executeQuery("CREATE INDEX IDX_62A8A7A7C79C849A ON registration (subsite_id);");
+        }
 
-        $conn->executeQuery("ALTER TABLE user_app ADD subsite_id INT DEFAULT NULL;");
-        $conn->executeQuery("ALTER TABLE user_app ADD CONSTRAINT FK_22781144C79C849A FOREIGN KEY (subsite_id) REFERENCES subsite (id) NOT DEFERRABLE INITIALLY IMMEDIATE;");
-        $conn->executeQuery("CREATE INDEX IDX_22781144C79C849A ON user_app (subsite_id);");
+        if(!__column_exists('user_app', 'subsite_id')){
+            $conn->executeQuery("ALTER TABLE user_app ADD subsite_id INT DEFAULT NULL;");
+            $conn->executeQuery("ALTER TABLE user_app ADD CONSTRAINT FK_22781144C79C849A FOREIGN KEY (subsite_id) REFERENCES subsite (id) NOT DEFERRABLE INITIALLY IMMEDIATE;");
+            $conn->executeQuery("CREATE INDEX IDX_22781144C79C849A ON user_app (subsite_id);");
+        }
     },
 
     'remove subsite slug column' => function () use($conn) {
+        if(!__column_exists('subsite', 'slug')){
+            return true;
+        }
+        
         $conn->executeQuery("ALTER TABLE subsite DROP COLUMN slug;");
     },
 
     'add subsite verified_seals column' => function () use($conn) {
+        if(__column_exists('subsite', 'verified_seals')){
+            return true;
+        }
         $conn->executeQuery("ALTER TABLE subsite ADD verified_seals VARCHAR(512) DEFAULT '[]';");
     },
     'update entities last_update_timestamp with user last log timestamp' => function () use($conn,$app) {
@@ -344,6 +381,9 @@ return [
     },
 
     'Created owner seal relation field' => function () use($conn) {
+        if(__column_exists('seal_relation', 'owner_id')){
+            return true;
+        }
         $conn->executeQuery("ALTER TABLE seal_relation ADD COLUMN owner_id INTEGER;");
         $agent_id = $conn->fetchColumn("select profile_id
                     from usr
@@ -388,10 +428,16 @@ return [
     },
 
     'Add field for maximum size from registration field configuration' => function () use($conn) {
+        if(__column_exists('registration_field_configuration', 'max_size')){
+            return true;
+        }
         $conn->executeQuery("ALTER TABLE registration_field_configuration ADD COLUMN max_size text;");
     },
 
     'Add notification type for compliant and suggestion messages' => function () use($conn) {
+        if(__table_exists('notification_meta')){
+            return true;
+        }
         $conn->executeQuery("CREATE TABLE notification_meta (id INT NOT NULL, object_id INT DEFAULT NULL, key VARCHAR(255) NOT NULL, value TEXT DEFAULT NULL, PRIMARY KEY(id));");
         $conn->executeQuery("CREATE SEQUENCE notification_meta_id_seq INCREMENT BY 1 MINVALUE 1 START 1;");
         $conn->executeQuery("ALTER TABLE notification_meta ADD CONSTRAINT notification_meta_fk FOREIGN KEY (object_id) REFERENCES notification (id) NOT DEFERRABLE INITIALLY IMMEDIATE;");
